FPM80 implemented addProductToCart method

+++ src/main/java/com/ecomerceApi/Priscila/request_responseModels/CartRequest.java
@@ -17,8 +17,9 @@ public class CartRequest {
     @NotEmpty
     @NonNull
     @Min(value = 1)
-    private Long quantity;
+    private int quantity;
 
 }
 
  */
+
